Refactor AI generators and lore store modules
