+++ src/lib/ai/narrativeGenerator.prompt.types.ts
@@ -1,0 +1,4 @@
+export interface NarrativeStaticContentCache {
+  itemAcquisitionInstructions?: string;
+  toneSettings?: Map<string, string>;
+}
