+++ src/lib/test-utils/mockStoreFactories/lore.ts
@@ -1,0 +1,49 @@
+import type { LoreStore } from './types';
+
+export function createMockLoreStore(
+  overrides?: Partial<LoreStore>
+): LoreStore {
+  return {
+    facts: {},
+    entities: {},
+    factHistory: {},
+    mergeAuditLog: [],
+    loreUsage: {},
+    loreUsageEvents: [],
+    currentEntityId: null,
+    error: null,
+    loading: false,
+    create: jest.fn(),
+    update: jest.fn(),
+    delete: jest.fn(),
+    setCurrent: jest.fn(),
+    getById: jest.fn(() => undefined),
+    getAll: jest.fn(() => []),
+    reset: jest.fn(),
+    setError: jest.fn(),
+    clearError: jest.fn(),
+    setLoading: jest.fn(),
+    addFact: jest.fn(),
+    getFacts: jest.fn(() => []),
+    clearFacts: jest.fn(),
+    cleanupOldFacts: jest.fn(),
+    compactFactHistory: jest.fn(),
+    getFactsCount: jest.fn(() => 0),
+    updateFact: jest.fn(),
+    deleteFact: jest.fn(),
+    validateFactUniqueness: jest.fn(() => true),
+    findSimilarFacts: jest.fn(() => []),
+    searchFacts: jest.fn(() => []),
+    exportFacts: jest.fn(() => '[]'),
+    importFacts: jest.fn(),
+    getFactHistory: jest.fn(() => []),
+    validateFact: jest.fn(() => true),
+    validateKey: jest.fn(() => true),
+    getLoreContext: jest.fn(() => ({ facts: [], factCount: 0, factIds: [] })),
+    addStructuredLore: jest.fn(),
+    recordLoreUsage: jest.fn(),
+    recordLoreMentions: jest.fn(),
+    clearLoreUsage: jest.fn(),
+    ...overrides,
+  } as LoreStore;
+}
