+++ src/lib/test-utils/mockStoreFactories/session.ts
@@ -1,0 +1,57 @@
+import type { SessionStore } from '@/types/game.types';
+
+export function createMockSessionStore(
+  overrides?: Partial<SessionStore>
+): SessionStore {
+  return {
+    id: null,
+    status: 'idle',
+    currentSceneId: null,
+    playerChoices: [],
+    error: null,
+    worldId: null,
+    characterId: null,
+    savedSessions: {},
+    sessionLifecycle: {},
+    templateHistory: [],
+    autoSave: {
+      enabled: true,
+      status: 'idle',
+      lastSaveTime: null,
+      errorMessage: null,
+      totalSaves: 0,
+    },
+    onboardingCompleted: false,
+    narrativeHeight: 400,
+    initializeSession: jest.fn().mockResolvedValue(undefined),
+    endSession: jest.fn(),
+    setStatus: jest.fn(),
+    setError: jest.fn(),
+    setPlayerChoices: jest.fn(),
+    selectChoice: jest.fn(),
+    clearPlayerChoices: jest.fn(),
+    setCurrentScene: jest.fn(),
+    pauseSession: jest.fn(),
+    resumeSession: jest.fn(),
+    setSessionId: jest.fn(),
+    setCharacterId: jest.fn(),
+    getSavedSession: jest.fn(() => undefined),
+    resumeSavedSession: jest.fn(() => false),
+    deleteSavedSession: jest.fn(),
+    updateSavedSessionNarrativeCount: jest.fn(),
+    fixExistingSessionNarrativeCounts: jest.fn().mockResolvedValue(undefined),
+    upsertSessionLifecycle: jest.fn(),
+    setSessionLifecycleStatus: jest.fn(),
+    getSessionLifecycle: jest.fn(() => undefined),
+    addTemplateToHistory: jest.fn(),
+    getTemplateHistory: jest.fn(() => []),
+    clearTemplateHistory: jest.fn(),
+    setAutoSaveEnabled: jest.fn(),
+    updateAutoSaveStatus: jest.fn(),
+    recordAutoSave: jest.fn(),
+    setOnboardingCompleted: jest.fn(),
+    isFirstTimeUser: jest.fn(() => true),
+    shouldShowOnboarding: jest.fn(() => false),
+    ...overrides,
+  } as SessionStore;
+}
