+++ src/state/loreStore.actions.types.ts
@@ -1,0 +1,9 @@
+import type { LoreStore } from './loreStore';
+
+export type SetState = (
+  partial:
+    | Partial<LoreStore>
+    | ((state: LoreStore) => Partial<LoreStore>)
+) => void;
+
+export type GetState = () => LoreStore;
